storyboard version1 .png
Storyboard:
US 01.08.01
As an owner, I want to ignore certain experimenters results.  ？how to draw

US 01.08.01
As an owner, I want to ignore certain experimenters results.  how to draw

US 01.05.01
As an experimenter, I want to be able to execute trials for an experiment and upload them to the experime
Connect this with QR?
Should it be specified to subscribed experiments?


<US 03.01.01
As an experimenter, I want to be able to generate QR codes that I can print for a specific experiment and trial result (for instance PASS for a binomial trial I subscribed to).

US 03.02.01

As an experimenter, I want to be able scan QR codes to indicate success or failure, or increment counts for trials I subscribed to.>

Are they the same QR code?



US 01.04.01
As an owner or experimenter, I want to subscribe to an experiment to participate in it.  ?   can owner subscribe his own experiment?

US 03.03.01
As an experimenter, I want to be able to register an arbitrary bar code (such as one off of your favourite book) to act a specific experiment result for a trial.
What’s this?

US 05.02.01
As an experimenter, I want search results to show each experiment with its description, owner username, and status.
What does status mean? Published/end/unpub/subs/owned?

No login; generate id automatically:
https://eclass.srv.ualberta.ca/mod/forum/discuss.php?d=1659865

Do we need to draw what list/page looks like after click/update/ add a new exp or reply or question?

Can owner search?

Minimum trials?